refactor: group root app setup, extract coach session + stage HUD parts
Two low-risk cleanups (all type-check + build verified):

Root layout — group the app-setup modules out of src/ root into src/app/
(router, i18n, directives, seo) and move PwaPrompt.vue into shell/. Root now
holds only entrypoints (App.vue, main.ts, style.css, vite-env). i18n's locale
glob/static paths were repointed to ../locales and the build still emits every
locale chunk. (Left the ui/ shadcn-style component names unprefixed — that's the
reka-ui/shadcn convention; renaming would fight the pattern.)

ViewerStage (920 -> 711):
- useCoachSession — coach tools + per-round board wiring (drawings/overrides/
  grenades + add/move/remove); coachMode stays owned by the stage and is
  injected (the demo-switch reset needs it before this runs).
- StageBadges — the transient status pills (comment/coach hints, follow badge,
  export-error banner).
- StageContextMenu — the right-click menu content + the round-color/outcome
  helpers it alone used.

type-check and production build pass.
