+++ apps/app/src/viewer/analysis/heatmap/HeatmapView.vue
@@ -13,7 +13,7 @@ import UiSelect from '@/ui/UiSelect.vue'
 import RoundTimeRange from '@/viewer/analysis/shared/RoundTimeRange.vue'
 import RoundStrip from '@/viewer/analysis/shared/RoundStrip.vue'
 import { freezeSeconds, maxLiveRoundTime } from '@/viewer/analysis/shared/roundTime'
-import { useI18n } from '@/i18n'
+import { useI18n } from '@/app/i18n'
 
 const { t } = useI18n()
