+++ apps/app/src/viewer/analysis/shared/RoundStrip.vue
@@ -4,7 +4,7 @@ import type { Round, Side } from '@/viewer/domain/schema'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
 import { roundDisplayLabels } from '@/viewer/domain/rounds'
 import UiIcon from '@/ui/UiIcon.vue'
-import { useI18n } from '@/i18n'
+import { useI18n } from '@/app/i18n'
 
 const { t } = useI18n()
