+++ apps/app/src/viewer/player/ViewerRoster.vue
@@ -3,7 +3,7 @@ import { computed } from 'vue'
 import type { PlayerMeta, PlayerState, Side } from '@/viewer/domain/schema'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
 import { weaponIconPath } from '@/viewer/domain/weaponIcons'
-import { useI18n } from '@/i18n'
+import { useI18n } from '@/app/i18n'
 
 const { t } = useI18n()
